+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### API changes
 
+* Replaced `SpaceEx.Stream.stream_fn/2` with `SpaceEx.Stream.with_get_fn/1`.
+  * `... |> Stream.stream_fn` is now `... |> Stream.stream |> Stream.with_get_fn`
 * Removed the `conn` argument from `SpaceEx.Event.create/1`.
   * It can derive `conn` from `expr`.
   * This also allows for expressions to be pipelined in.